requirements.txt
aiogram==3.22.0
alembic==1.17.0
asyncpg==0.30.0
mypy==1.18.2
psycopg2==2.9.11
python-dotenv==1.2.1
ruff==0.14.5
sqlalchemy==2.0.44
